pyproject.toml
[tool.poetry]
name = "pasteme"
version = "0.1.0"
description = ""
authors = ["lim-mil <limyel@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.8"
starlette = {extras = ["full"], version = "^0.14.2"}
uvicorn = "^0.13.3"
peewee = "^3.14.1"
peewee-async = "^0.7.1"
aiomysql = "^0.0.21"
cryptography = "^3.4.4"
pyjwt = "^2.0.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
